[web] Move ui files to lib/ (#30051)
