+++ ci/licenses_golden/licenses_flutter
@@ -493,6 +493,22 @@ FILE: ../../../flutter/lib/ui/window/viewport_metrics.cc
 FILE: ../../../flutter/lib/ui/window/viewport_metrics.h
 FILE: ../../../flutter/lib/ui/window/window.cc
 FILE: ../../../flutter/lib/ui/window/window.h
+FILE: ../../../flutter/lib/web_ui/lib/annotations.dart
+FILE: ../../../flutter/lib/web_ui/lib/canvas.dart
+FILE: ../../../flutter/lib/web_ui/lib/channel_buffers.dart
+FILE: ../../../flutter/lib/web_ui/lib/compositing.dart
+FILE: ../../../flutter/lib/web_ui/lib/geometry.dart
+FILE: ../../../flutter/lib/web_ui/lib/hash_codes.dart
+FILE: ../../../flutter/lib/web_ui/lib/initialization.dart
+FILE: ../../../flutter/lib/web_ui/lib/key.dart
+FILE: ../../../flutter/lib/web_ui/lib/lerp.dart
+FILE: ../../../flutter/lib/web_ui/lib/natives.dart
+FILE: ../../../flutter/lib/web_ui/lib/painting.dart
+FILE: ../../../flutter/lib/web_ui/lib/path.dart
+FILE: ../../../flutter/lib/web_ui/lib/path_metrics.dart
+FILE: ../../../flutter/lib/web_ui/lib/platform_dispatcher.dart
+FILE: ../../../flutter/lib/web_ui/lib/pointer.dart
+FILE: ../../../flutter/lib/web_ui/lib/semantics.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/alarm_clock.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/assets.dart
@@ -641,26 +657,10 @@ FILE: ../../../flutter/lib/web_ui/lib/src/engine/util.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/validators.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/vector_math.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/window.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/annotations.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/canvas.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/channel_buffers.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/compositing.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/geometry.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/hash_codes.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/initialization.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/key.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/lerp.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/natives.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/painting.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/path.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/path_metrics.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/platform_dispatcher.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/pointer.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/semantics.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/text.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/tile_mode.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/ui/window.dart
+FILE: ../../../flutter/lib/web_ui/lib/text.dart
+FILE: ../../../flutter/lib/web_ui/lib/tile_mode.dart
 FILE: ../../../flutter/lib/web_ui/lib/ui.dart
+FILE: ../../../flutter/lib/web_ui/lib/window.dart
 FILE: ../../../flutter/lib/web_ui/tool/unicode_sync_script.dart
 FILE: ../../../flutter/runtime/dart_isolate.cc
 FILE: ../../../flutter/runtime/dart_isolate.h
