+++ lib/ui/semantics/string_attribute.h
@@ -18,7 +18,7 @@ using StringAttributes = std::vector<StringAttributePtr>;
 // When adding a new StringAttributeType, the classes in these file must be
 // updated as well.
 //  * engine/src/flutter/lib/ui/semantics.dart
-//  * engine/src/flutter/lib/web_ui/lib/src/ui/semantics.dart
+//  * engine/src/flutter/lib/web_ui/lib/semantics.dart
 //  * engine/src/flutter/shell/platform/android/io/flutter/view/AccessibilityBridge.java
 //  * engine/src/flutter/lib/web_ui/test/engine/semantics/semantics_api_test.dart
 //  * engine/src/flutter/testing/dart/semantics_test.dart
