+++ lib/web_ui/lib/path_metrics.dart
