+++ web_sdk/BUILD.gn
@@ -14,11 +14,9 @@ dart_sdk_package_config = "//third_party/dart/.dart_tool/package_config.json"
 web_ui_sources = exec_script("//third_party/dart/tools/list_dart_files.py",
                              [
                                "absolute",
-                               rebase_path("//flutter/lib/web_ui/lib/src/ui"),
+                               rebase_path("//flutter/lib/web_ui/lib"),
                              ],
                              "list lines")
-
-web_ui_sources += [ "//flutter/lib/web_ui/lib/ui.dart" ]
 
 web_engine_sources =
     exec_script("//third_party/dart/tools/list_dart_files.py",
